docs(multiselect): add example with hooks
